lab-agent/requirements.txt
dotenv
azure-ai-projects
azure-identity
azure-ai-inference
azure-monitor-opentelemetry